Data.Drasil.ExternalLibraries.ODELibraries

Description

Define and collect information about ODEs and ODE solvers from various libraries.

SciPy Library (Python)

scipyODEPckg :: ODELibPckg

SciPy ODE library package.

scipyODESymbols :: [DefinedQuantityDict]

Collects variables needed for SciPy's ODEs as DefinedQuantityDicts.

Oslo Library (C#)

osloPckg :: ODELibPckg

Oslo ODE library package.

osloSymbols :: [DefinedQuantityDict]

Collects variables needed for Oslo's ODEs as DefinedQuantityDicts.

arrayVecDepVar :: ODEInfo -> CodeVarChunk

Oslo's Vector type behaves like an array, so needs to be represented as one or else will hit type errors in GOOL.

Apache Commons (Java)

apacheODEPckg :: ODELibPckg

Apache Commons ODE library package.

apacheODESymbols :: [DefinedQuantityDict]

Collects variables needed for Apache's ODEs as DefinedQuantityDicts.

Odeint (C++)

odeintPckg :: ODELibPckg

odeint ODE library package.

odeintSymbols :: [DefinedQuantityDict]

Collects variables needed for odeint's ODEs as DefinedQuantityDicts.

diffCodeChunk :: CodeVarChunk -> CodeVarChunk

Change in X chunk constructor (where X is a given argument).