Language.Drasil.Choices

Description

Defines the design language for SCS.

Documentation

data Choices

The instruction indicates how the generated program should be written down. Full details of Choices documentation https://github.com/JacquesCarette/Drasil/wiki/The-Code-Generator

Constructors

Choices

Fields

lang :: [Lang]

Target languages. Choosing multiple means program will be generated in multiple languages.

architecture :: Architecture

Architecture of the program, include modularity and implementation type

dataInfo :: DataInfo

Data structure and represent

maps :: Maps

Maps for 'Drasil concepts' to 'code concepts' or Space to a 'CodeType

optFeats :: OptionalFeatures

Setting for Softifacts that can be added to the program or left it out

srsConstraints :: Constraints

Constraint violation behaviour. Exception or Warning.

extLibs :: [ExtLib]

List of external libraries what to utilize

icNames :: InternalConcept -> Name

Function to get modifiable function names

folderVal :: Int

Number of folders to go up in order to obtain the image

data Architecture

Architecture of a program

Constructors

Archt

Fields

modularity :: Modularity

How the program should be modularized.

impType :: ImplementationType

Implementation type, program or library.

makeArchit :: Modularity -> ImplementationType -> Architecture

Constructor to create a Architecture

data DataInfo

Data of a program - how information should be encoded.

Constructors

DataInfo

Fields

inputStructure :: Structure

Structure of inputs (bundled or not).

constStructure :: ConstantStructure

Structure of constants (inlined or bundled or not, or stored with inputs).

constRepr :: ConstantRepr

Representation of constants (as variables or as constants).

makeData :: Structure -> ConstantStructure -> ConstantRepr -> DataInfo

Constructor to create a DataInfo

data Maps

Maps for Concepts and Space

Constructors

Maps

Fields

conceptMatch :: ConceptMatchMap

Map of UIDs for Drasil concepts to code concepts. Matching a UID to a code concept means the code concept should be used instead of the chunk associated with the UID.

spaceMatch :: SpaceMatch

Map of Spaces to CodeTypes Matching a Space to a CodeType means values of the Space should have that CodeType in the generated code.

makeMaps :: ConceptMatchMap -> SpaceMatch -> Maps

Constructor to create a Maps

spaceToCodeType :: Space -> [CodeType]

Default mapping between Space and CodeType.

data Constraints

SRS Constraints

Constructors

Constraints

Fields

onSfwrConstraint :: ConstraintBehaviour

onPhysConstraint :: ConstraintBehaviour

makeConstraints :: ConstraintBehaviour -> ConstraintBehaviour -> Constraints

Constructor to create a Constraints

data ODE

All Information needed to solve an ODE

Constructors

ODE

Fields

odeInfo :: [ODEInfo]

ODE information.

odeLib :: [ODELibPckg]

Preferentially-ordered list ODE libraries to try.

makeODE :: [ODEInfo] -> [ODELibPckg] -> ODE

Constructor to create an ODE

data DocConfig

Configuration for Doxygen documentation

Constructors

DocConfig

Fields

comments :: [Comments]

Turns Doxygen comments for different code structures on or off.

doxVerbosity :: Verbosity

Standard output from running Doxygen: verbose or quiet?

dates :: Visibility

Turns date field on or off in the generated module-level Doxygen comments.

makeDocConfig :: [Comments] -> Verbosity -> Visibility -> DocConfig

Constructor to create a DocConfig

data LogConfig

Log Configuration

Constructors

LogConfig

Fields

logging :: [Logging]

Turns different forms of logging on or off.

logFile :: FilePath

Name of log file.

makeLogConfig :: [Logging] -> FilePath -> LogConfig

Constructor to create a LogConfig

data OptionalFeatures

Constructors

OptFeats

Fields

docConfig :: DocConfig

logConfig :: LogConfig

auxFiles :: [AuxFile]

Turns generation of different auxiliary (non-source-code) files on or off.

makeOptFeats :: DocConfig -> LogConfig -> [AuxFile] -> OptionalFeatures

Constructor to create a OptionalFeatures

newtype ExtLib

External Library Options

Constructors

Math ODE

data Modularity

Modularity of a program.

Constructors

Modular
Different modules. For controller, input, calculations, output.

Unmodular
All generated code is in one module/file.

data Structure

Variable structure options.

Constructors

Unbundled
Individual variables

Bundled
Variables bundled in a class

data ConstantStructure

Constants options.

Constructors

Inline
Inline values for constants.

WithInputs
Store constants with inputs.

Store Structure
Store constants separately from inputs, whether bundled or unbundled.

data ConstantRepr

Options for representing constants in a program.

Constructors

Var
Constants represented as regular variables.

Const
Use target language's mechanism for defining constants.

type ConceptMatchMap = Map UID [CodeConcept]

Specifies matches between chunks and CodeConcepts, meaning the target language's pre-existing definition of the concept should be used instead of defining a new variable for the concept in the generated code. [CodeConcept] is preferentially-ordered, generator concretizes a ConceptMatchMap to a MatchedConceptMap by checking user's other choices.

type MatchedConceptMap = Map UID CodeConcept

Concrete version of ConceptMatchMap dependent on user choices.

data CodeConcept

Code concepts. For now, just pi.

Constructors

Pi

matchConcepts :: HasUID c => [(c, [CodeConcept])] -> ConceptMatchMap

Builds a ConceptMatchMap from an association list of chunks and CodeConcepts.

type SpaceMatch = Space -> [CodeType]

Specifies which CodeType should be used to represent each mathematical Space. [CodeType] is preferentially-ordered, first CodeType that does not conflict with other choices will be selected.

matchSpaces :: [(Space, [CodeType])] -> SpaceMatch

Builds a SpaceMatch from an association list of Spaces and CodeTypes.

data ImplementationType

Program implementation options.

Constructors

Library
Generated code does not include Controller.

Program
Generated code includes Controller.

data ConstraintBehaviour

Constraint behaviour options within program.

Constructors

Warning
Print warning when constraint violated.

Exception
Throw exception when constraint violated.

data Comments

Comment implementation options.

Constructors

CommentFunc
Function/method-level comments.

CommentClass
Class-level comments.

CommentMod
File/Module-level comments.

data Verbosity

Doxygen file verbosity options.

Constructors

Verbose

Quiet

data Visibility

Doxygen date-field visibility options.

Constructors

Show

Hide

data Logging

Logging options for function calls and variable assignments. Eq instances required for Logging and Comments because generator needs to check membership of these elements in lists

Constructors

LogFunc
Log messages generated for function calls.

LogVar
Log messages generated for variable assignments.

data AuxFile

Currently we only support two kind of auxiliary files: sample input file, readme. To generate a sample input file compatible with the generated program, FilePath is the path to the user-provided file containing a sample set of input data.

Constructors

SampleInput FilePath

ReadME

getSampleData :: Choices -> Maybe FilePath

Gets the file path to a sample input data set from a Choices structure, if the user chose to generate a sample input file.

hasSampleInput :: [AuxFile] -> Bool

Predicate that returns true if the list of AuxFiles includes a SampleInput.

defaultChoices :: Choices

Default choices to be used as the base from which design specifications can be built.

choicesSent :: Choices -> [Sentence]

Renders Choices as Sentences.

showChs :: RenderChoices a => a -> Sentence

data InternalConcept

Data type of internal concepts

Constructors

InputConstraintsFn

InputConstraintsMod

WriteOutput

DerivedValuesFn

DerivedValuesMod

GetInput

InputParameters

InputFormat

OutputFormat

Calculations

Constants