Language.Drasil.Chunk.Code

Description

Defines chunk types for use in code generation.

Documentation

class CodeIdea c where

A CodeIdea must include some code and its name.

Methods

codeName :: c -> String

Name of the idea.

codeChunk :: c -> CodeChunk

Code chunk associated with the idea.

data CodeChunk

Basic chunk representation in the code generation context. Contains a QuantityDict and the kind of code (variable or function).

Constructors

CodeC

Fields

_qc :: QuantityDict

kind :: VarOrFunc

data CodeVarChunk

Chunk representing a variable. The obv field represents the object containing this variable, if it is an object field.

Constructors

CodeVC

Fields

_ccv :: CodeChunk

_obv :: Maybe CodeChunk

newtype CodeFuncChunk

Chunk representing a function.

Constructors

CodeFC

Fields

_ccf :: CodeChunk

data VarOrFunc

Details if a piece of code is meant to be a variable or a function.

Constructors

Var

Func

obv :: Lens' CodeVarChunk (Maybe CodeChunk)

quantvar :: (Quantity c, MayHaveUnit c) => c -> CodeVarChunk

Construct a CodeVarChunk from a Quantity.

quantfunc :: (Quantity c, MayHaveUnit c) => c -> CodeFuncChunk

Construct a CodeFuncChunk from a Quantity.

ccObjVar :: CodeVarChunk -> CodeVarChunk -> CodeVarChunk

Combine an Object-type CodeChunk with another CodeChunk to create a new CodeChunk which represents a field of the first. ex. ccObjVar obj f = obj.f.

codevars :: CodeExpr -> ChunkDB -> [CodeVarChunk]

Get a list of CodeChunks from an equation.

codevars' :: CodeExpr -> ChunkDB -> [CodeVarChunk]

Get a list of CodeChunks from an equation (no functions).

funcResolve :: ChunkDB -> UID -> CodeFuncChunk

Make a CodeFuncChunk from a UID in the ChunkDB.

varResolve :: ChunkDB -> UID -> CodeVarChunk

Make a CodeVarChunk from a UID in the ChunkDB.

listToArray :: CodeVarChunk -> CodeVarChunk

programName :: CommonIdea c => c -> String

Convert the program name to an abbreviated String without any special characters.

funcPrefix :: String

Used when a function name needs to be distinguishable from a variable name.

class CodeIdea c => DefiningCodeExpr c where

A DefiningCodeExpr must have it's underlying chunk defined in the CodeExpr language.

Methods

codeExpr :: Lens' c CodeExpr

Orphan instances

CodeIdea CodeFuncChunk
Finds the code name and CodeChunk within a CodeFuncChunk.

Instance details

Methods

codeName :: CodeFuncChunk -> String

codeChunk :: CodeFuncChunk -> CodeChunk

CodeIdea CodeVarChunk
Finds the code name and CodeChunk within a CodeVarChunk.

Instance details

Methods

codeName :: CodeVarChunk -> String

codeChunk :: CodeVarChunk -> CodeChunk

CodeIdea CodeChunk
Finds the code name of a CodeChunk.

Instance details

Methods

codeName :: CodeChunk -> String

codeChunk :: CodeChunk -> CodeChunk