Language.Drasil.Chunk.ConstraintMap

Documentation

type ConstraintCEMap = Map UID [ConstraintCE]

Constraints map. Contains all Constraints.

type ConstraintCE = Constraint CodeExpr

Type synonym for 'Constraint CodeExpr'

constraintMap :: (HasUID c, Constrained c) => [c] -> ConstraintCEMap

Creates a map from UID to Constraints for constrained chunks.

physLookup :: HasUID q => ConstraintCEMap -> q -> (q, [ConstraintCE])

Returns a pair of a chunk and its physical constraints.

sfwrLookup :: HasUID q => ConstraintCEMap -> q -> (q, [ConstraintCE])

Returns a pair of a chunk and its software constraints.