Language.Drasil.Code

Description

Re-export code-related smart constructors for external code writing and generation.

Documentation

makeCode :: [FileData] -> [AuxData] -> Code

Makes code from FileData (FilePaths with module data) and AuxData (FilePaths with auxiliary document information).

createCodeFiles :: Code -> IO ()

Creates the requested Code by producing files.

generator :: Lang -> String -> [Expr] -> Choices -> CodeSpec -> DrasilState

Initializes the generator's DrasilState. Space parameter is a string representing the date. \['Expr'\] parameter is the sample input values provided by the user.

generateCode :: (OOProg progRepr, PackageSym packRepr) => Lang -> (progRepr (Program progRepr) -> ProgData) -> (packRepr (Package packRepr) -> PackData) -> DrasilState -> IO ()

Generates a package with the given DrasilState. The passed un-representation functions determine which target language the package will be generated in.

generateCodeProc :: (ProcProg progRepr, PackageSym packRepr) => Lang -> (progRepr (Program progRepr) -> ProgData) -> (packRepr (Package packRepr) -> PackData) -> DrasilState -> IO ()

Generates a package with the given DrasilState. The passed un-representation functions determine which target language the package will be generated in.

readWithDataDesc :: FilePath -> DataDesc' -> IO [Expr]

Reads data from a file and converts the values to Exprs. The file must be formatted according to the DataDesc' passed as a parameter.

sampleInputDD :: [CodeVarChunk] -> DataDesc'

Defines the DataDesc for the file containing a sample data set, which a user must supply if they want to generate a sample input file.

data Choices

The instruction indicates how the generated program should be written down. Full details of Choices documentation https://github.com/JacquesCarette/Drasil/wiki/The-Code-Generator

Constructors

Choices

Fields

lang :: [Lang]

Target languages. Choosing multiple means program will be generated in multiple languages.

architecture :: Architecture

Architecture of the program, include modularity and implementation type

dataInfo :: DataInfo

Data structure and represent

maps :: Maps

Maps for 'Drasil concepts' to 'code concepts' or Space to a 'CodeType

optFeats :: OptionalFeatures

Setting for Softifacts that can be added to the program or left it out

srsConstraints :: Constraints

Constraint violation behaviour. Exception or Warning.

extLibs :: [ExtLib]

List of external libraries what to utilize

icNames :: InternalConcept -> Name

Function to get modifiable function names

folderVal :: Int

Number of folders to go up in order to obtain the image

data Comments

Comment implementation options.

Constructors

CommentFunc
Function/method-level comments.

CommentClass
Class-level comments.

CommentMod
File/Module-level comments.

data Verbosity

Doxygen file verbosity options.

Constructors

Verbose

Quiet

data ConstraintBehaviour

Constraint behaviour options within program.

Constructors

Warning
Print warning when constraint violated.

Exception
Throw exception when constraint violated.

makeArchit :: Modularity -> ImplementationType -> Architecture

Constructor to create a Architecture

data Architecture

Architecture of a program

Constructors

Archt

Fields

modularity :: Modularity

How the program should be modularized.

impType :: ImplementationType

Implementation type, program or library.

data DataInfo

Data of a program - how information should be encoded.

Constructors

DataInfo

Fields

inputStructure :: Structure

Structure of inputs (bundled or not).

constStructure :: ConstantStructure

Structure of constants (inlined or bundled or not, or stored with inputs).

constRepr :: ConstantRepr

Representation of constants (as variables or as constants).

makeData :: Structure -> ConstantStructure -> ConstantRepr -> DataInfo

Constructor to create a DataInfo

data Maps

Maps for Concepts and Space

Constructors

Maps

Fields

conceptMatch :: ConceptMatchMap

Map of UIDs for Drasil concepts to code concepts. Matching a UID to a code concept means the code concept should be used instead of the chunk associated with the UID.

spaceMatch :: SpaceMatch

Map of Spaces to CodeTypes Matching a Space to a CodeType means values of the Space should have that CodeType in the generated code.

makeMaps :: ConceptMatchMap -> SpaceMatch -> Maps

Constructor to create a Maps

spaceToCodeType :: Space -> [CodeType]

Default mapping between Space and CodeType.

makeConstraints :: ConstraintBehaviour -> ConstraintBehaviour -> Constraints

Constructor to create a Constraints

makeODE :: [ODEInfo] -> [ODELibPckg] -> ODE

Constructor to create an ODE

makeDocConfig :: [Comments] -> Verbosity -> Visibility -> DocConfig

Constructor to create a DocConfig

makeLogConfig :: [Logging] -> FilePath -> LogConfig

Constructor to create a LogConfig

data LogConfig

Log Configuration

Constructors

LogConfig

Fields

logging :: [Logging]

Turns different forms of logging on or off.

logFile :: FilePath

Name of log file.

data OptionalFeatures

Constructors

OptFeats

Fields

docConfig :: DocConfig

logConfig :: LogConfig

auxFiles :: [AuxFile]

Turns generation of different auxiliary (non-source-code) files on or off.

makeOptFeats :: DocConfig -> LogConfig -> [AuxFile] -> OptionalFeatures

Constructor to create a OptionalFeatures

newtype ExtLib

External Library Options

Constructors

Math ODE

data ImplementationType

Program implementation options.

Constructors

Library
Generated code does not include Controller.

Program
Generated code includes Controller.

data Logging

Logging options for function calls and variable assignments. Eq instances required for Logging and Comments because generator needs to check membership of these elements in lists

Constructors

LogFunc
Log messages generated for function calls.

LogVar
Log messages generated for variable assignments.

data Modularity

Modularity of a program.

Constructors

Modular
Different modules. For controller, input, calculations, output.

Unmodular
All generated code is in one module/file.

data Structure

Variable structure options.

Constructors

Unbundled
Individual variables

Bundled
Variables bundled in a class

data ConstantStructure

Constants options.

Constructors

Inline
Inline values for constants.

WithInputs
Store constants with inputs.

Store Structure
Store constants separately from inputs, whether bundled or unbundled.

data ConstantRepr

Options for representing constants in a program.

Constructors

Var
Constants represented as regular variables.

Const
Use target language's mechanism for defining constants.

data CodeConcept

Code concepts. For now, just pi.

Constructors

Pi

matchConcepts :: HasUID c => [(c, [CodeConcept])] -> ConceptMatchMap

Builds a ConceptMatchMap from an association list of chunks and CodeConcepts.

type SpaceMatch = Space -> [CodeType]

Specifies which CodeType should be used to represent each mathematical Space. [CodeType] is preferentially-ordered, first CodeType that does not conflict with other choices will be selected.

matchSpaces :: [(Space, [CodeType])] -> SpaceMatch

Builds a SpaceMatch from an association list of Spaces and CodeTypes.

data AuxFile

Currently we only support two kind of auxiliary files: sample input file, readme. To generate a sample input file compatible with the generated program, FilePath is the path to the user-provided file containing a sample set of input data.

Constructors

SampleInput FilePath

ReadME

getSampleData :: Choices -> Maybe FilePath

Gets the file path to a sample input data set from a Choices structure, if the user chose to generate a sample input file.

data Visibility

Doxygen date-field visibility options.

Constructors

Show

Hide

defaultChoices :: Choices

Default choices to be used as the base from which design specifications can be built.

data CodeSpec

New Code Specification. Holds system information and a reference to OldCodeSpec.

Constructors

CS

Fields

_system' :: System

_oldCode :: OldCodeSpec

data OldCodeSpec

Old Code specifications. Holds information needed to generate code.

Constructors

OldCodeSpec

Fields

_pName :: Name

Program name.

_authors :: People

Authors.

_inputs :: [Input]

All inputs.

_extInputs :: [Input]

Explicit inputs (values to be supplied by a file).

_derivedInputs :: [Derived]

Derived inputs (each calculated from explicit inputs in a single step).

_outputs :: [Output]

All outputs.

_configFiles :: [FilePath]

List of files that must be in same directory for running the executable.

_execOrder :: [Def]

Mathematical definitions, ordered so that they form a path from inputs to outputs.

_cMap :: ConstraintCEMap

Map from UIDs to constraints for all constrained chunks used in the problem.

_constants :: [Const]

List of all constants used in the problem.

_constMap :: ConstantMap

Map containing all constants used in the problem.

_mods :: [Mod]

Additional modules required in the generated code, which Drasil cannot yet automatically define.

_systemdb :: ChunkDB

The database of all chunks used in the problem.

codeSpec :: System -> Choices -> [Mod] -> CodeSpec

Creates a CodeSpec using the provided System, Choices, and Mods. The CodeSpec consists of the system information and a corresponding OldCodeSpec.

class HasOldCodeSpec c where

Minimal complete definition

oldCodeSpec

Methods

oldCodeSpec :: Lens' c OldCodeSpec

authorsO :: Lens' c People

cMapO :: Lens' c ConstraintCEMap

configFilesO :: Lens' c [FilePath]

constMapO :: Lens' c ConstantMap

constantsO :: Lens' c [Const]

derivedInputsO :: Lens' c [Derived]

execOrderO :: Lens' c [Def]

extInputsO :: Lens' c [Input]

inputsO :: Lens' c [Input]

modsO :: Lens' c [Mod]

outputsO :: Lens' c [Output]

pNameO :: Lens' c Name

systemdbO :: Lens' c ChunkDB

funcUID :: Func -> UID

Get a UID of a chunk corresponding to a Func.

asVC :: Func -> DefinedQuantityDict

Convert a Func to an implementation-stage DefinedQuantityDict representing the function.

($:=) :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> FuncStmt

Define an assignment statement.

data Mod

Holds module information.

Constructors

Mod Name Description [Import] [Class] [Func]

data StateVariable

State variables hold attach a VisibilityTag to a CodeVarChunk.

data Func

Holds a function definition or function data.

data FuncStmt where

Constructors

FAsg :: CodeVarChunk -> CodeExpr -> FuncStmt

FAsgIndex :: CodeVarChunk -> Integer -> CodeExpr -> FuncStmt

FFor :: CodeVarChunk -> CodeExpr -> CodeExpr -> CodeExpr -> [FuncStmt] -> FuncStmt
For-loop; Variable, Start, Stop, Step, Body.

FForEach :: CodeVarChunk -> CodeExpr -> [FuncStmt] -> FuncStmt

FWhile :: CodeExpr -> [FuncStmt] -> FuncStmt

FCond :: CodeExpr -> [FuncStmt] -> [FuncStmt] -> FuncStmt

FRet :: CodeExpr -> FuncStmt

FThrow :: String -> FuncStmt

FTry :: [FuncStmt] -> [FuncStmt] -> FuncStmt

FContinue :: FuncStmt

FDecDef :: CodeVarChunk -> CodeExpr -> FuncStmt

FFuncDef :: CodeFuncChunk -> [ParameterChunk] -> [FuncStmt] -> FuncStmt

FVal :: CodeExpr -> FuncStmt

FMulti :: [FuncStmt] -> FuncStmt

FAppend :: CodeExpr -> CodeExpr -> FuncStmt

pubStateVar :: CodeVarChunk -> StateVariable

Define a public state variable based on the given CodeVarChunk.

privStateVar :: CodeVarChunk -> StateVariable

Define a private state variable based on the given CodeVarChunk.

fDecDef :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> FuncStmt

Define a declare-define statement.

ffor :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> [FuncStmt] -> FuncStmt

Define a for-loop. Quantity is for the iteration variable, CodeExpr is the upper bound at that variable (the variable will start with a value of 0). [FuncStmt] is for the loop body.

fforRange :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> CodeExpr -> CodeExpr -> [FuncStmt] -> FuncStmt

Define a for-loop. Quantity is for the iteration variable, and 3 CodeExprs for the start, stop, step numbers. [FuncStmt] is for the loop body.

funcData :: Name -> Description -> DataDesc -> Func

Define a function that reads data from a file, according to the given DataDesc.

funcDef :: (Quantity c, MayHaveUnit c, Concept c) => Name -> Description -> [c] -> Space -> Maybe Description -> [FuncStmt] -> Func

Define a function by providing the FuncStmts for its body. Other parameters are function name, description, list of parameters, space of the returned value, and description of the returned value.

packmod :: Name -> Description -> [Class] -> [Func] -> Mod

Define a Mod with the given Name, Description, Classes, and Functions.

junkLine :: Data

Constructor for junk data.

multiLine :: LinePattern -> Delim -> Data

Constructor for an unknown amount of lines of data.

repeated :: [DataItem] -> LinePattern

Constructor for data with a repeated pattern.

singleLine :: LinePattern -> Delim -> Data

Constructor for a single line of data.

singleton :: DataItem -> Data

Constructor for a single datum.

type ExternalLibrary = [StepGroup]

External library is a group of Steps

data Step

A step can be a call to an external library function or method.

data FunctionInterface

The first item in the Requires list should be where the function being called is defined.

data Argument

An argument may contain a named argument and argument information.

externalLib :: [StepGroup] -> ExternalLibrary

Specifies an external library.

choiceSteps :: [[Step]] -> StepGroup

To be used when there are multiple options for a group of consecutive steps, where a single use-case-specific factor decides which step group to use.

choiceStep :: [Step] -> StepGroup

To be used when there are multiple options for a single step, where a use-case-specific factor decides which step to use.

mandatoryStep :: Step -> StepGroup

Specifies a step which must exist in some form in every use case.

mandatorySteps :: [Step] -> StepGroup

Specifies multiple consecutive steps that all must exist in some form in every use case.

callStep :: FunctionInterface -> Step

libFunction :: Requires -> CodeFuncChunk -> [Argument] -> FunctionInterface

Specifies a call to an external library function.

libMethod :: Requires -> CodeVarChunk -> CodeFuncChunk -> [Argument] -> FunctionInterface

Specifies a call to an external library method.

libFunctionWithResult :: Requires -> CodeFuncChunk -> [Argument] -> CodeVarChunk -> FunctionInterface

Specifies a call to an external library function, where the result is assigned to a variable.

libMethodWithResult :: Requires -> CodeVarChunk -> CodeFuncChunk -> [Argument] -> CodeVarChunk -> FunctionInterface

Specifies a call to an external library method, where the result is assigned to a variable.

libConstructor :: Requires -> CodeFuncChunk -> [Argument] -> CodeVarChunk -> FunctionInterface

Specifies a call to an external library constructor, where the result is assigned to a variable.

libConstructorMultiReqs :: [Requires] -> CodeFuncChunk -> [Argument] -> CodeVarChunk -> FunctionInterface

Specifies a call to an external library function, where multiple modules from the external library are required, and the result is assigned to a variable.

constructAndReturn :: Requires -> CodeFuncChunk -> [Argument] -> FunctionInterface

Specifies a call to an external library constructor, where the result is returned.

lockedArg :: CodeExpr -> Argument

Specifies an argument that is not use-case-dependent.

lockedNamedArg :: NamedArgument -> CodeExpr -> Argument

Specifies a named argument that is not use-case-dependent.

inlineArg :: Space -> Argument

Specifies a use-case-dependent argument whose value can be inlined in the call.

inlineNamedArg :: NamedArgument -> Space -> Argument

Specifies a use-case-dependent named argument whose value can be inlined in the call.

preDefinedArg :: CodeVarChunk -> Argument

Specifies use-case-dependent argument whose value must be assigned to a variable before being passed in the call.

preDefinedNamedArg :: NamedArgument -> CodeVarChunk -> Argument

Specifies use-case-dependent named argument whose value must be assigned to a variable before being passed in the call.

functionArg :: CodeFuncChunk -> [Parameter] -> Step -> Argument

Specifies a function type argument, where the body consists of a single step.

customObjArg :: [Requires] -> Description -> CodeVarChunk -> CodeFuncChunk -> ClassInfo -> Argument

Specifies an argument that is an object of a class that must be defined in the calling program.

recordArg :: Requires -> CodeFuncChunk -> CodeVarChunk -> [CodeVarChunk] -> Argument

Specifies an argument that is an object of a class from the external library. The list of [CodeVarChunk] represents fields of the object that must be set in the calling program.

lockedParam :: CodeVarChunk -> Parameter

Specifies a use-case-independent parameter.

unnamedParam :: Space -> Parameter

Specifies a parameter whose name depends on the use case.

customClass :: [MethodInfo] -> ClassInfo

Specifies a class that must be implemented in the calling program.

implementation :: String -> [MethodInfo] -> ClassInfo

Specifies an implementation of an interface from the external library.

constructorInfo :: CodeFuncChunk -> [Parameter] -> [Step] -> MethodInfo

Specifies a constructor.

methodInfo :: CodeFuncChunk -> Description -> [Parameter] -> Description -> [Step] -> MethodInfo

Specifies a method.

methodInfoNoReturn :: CodeFuncChunk -> Description -> [Parameter] -> [Step] -> MethodInfo

Specifies a method that does not return anything.

appendCurrSol :: CodeExpr -> Step

Specifies a statement where a current solution is appended to a solution list.

populateSolList :: CodeVarChunk -> CodeVarChunk -> CodeVarChunk -> [Step]

Specifies a statement where a solution list is populated by iterating through a solution array.

assignArrayIndex :: Step

Specifies statements where every index of an array is assigned a value.

assignSolFromObj :: CodeVarChunk -> Step

Specifies a statement where a solution is assigned from the field of an object.

initSolListFromArray :: CodeVarChunk -> Step

Specifies a statement where a solution list is initialized with the first element of an array.

initSolListWithVal :: Step

Specifies a statement where a solution list is initialized with the first value.

solveAndPopulateWhile :: FunctionInterface -> CodeVarChunk -> CodeVarChunk -> FunctionInterface -> CodeVarChunk -> Step

A solve and populate loop. FunctionInterface for loop condition, CodeChunk for solution object, CodeChunk for independent var, FunctionInterface for solving, CodeChunk for soln array to populate with.

returnExprList :: Step

Specifies a statement where a list is returned, where each value of the list is explicitly defined.

fixedReturn :: CodeExpr -> Step

Specifies a use-case-independent statement that returns a fixed value.

fixedReturn' :: Step

Specifies a use-case-dependent statement that returns a non-fixed value.

initSolWithVal :: Step

Specifies a statement where a single solution is initialized with a value.

type ExternalLibraryCall = [StepGroupFill]

External library call holds a group of step groups.

data StepGroupFill

Holds a group of steps (StepFills). The Int is to "choose" from the options in ExternalLibrary.

Constructors

SGF Int [StepFill]

data StepFill

Mirrors ExternalLibrary's Step. A StepFill can be a call to an external library function or method.

Constructors

CallF FunctionIntFill

LoopF (NonEmpty FunctionIntFill) [CodeExpr] (NonEmpty StepFill)

StatementF [CodeVarChunk] [CodeExpr]

newtype FunctionIntFill

Mirrors ExternalLibrary's FunctionInterface.

Constructors

FIF [ArgumentFill]

data ArgumentFill

Mirrors ExternalLibrary's ArgumentInfo. Determines the context needed for an argument to work.

Constructors

UserDefinedArgF (Maybe NamedArgument) CodeExpr
For arguments that are completely dependent on use case.

BasicF CodeExpr
A basic function.

FnF [ParameterFill] StepFill
Fills in the names for the unnamed parameters.

ClassF [StateVariable] ClassInfoFill
List of CodeChunk for state variables.

RecordF [CodeExpr]
Fills in the field values.

data ParameterFill

Mirrors ExternalLibrary's Parameter.

Constructors

NameableParamF ParameterChunk

UserDefined ParameterChunk

data ClassInfoFill

Mirrors ExternalLibrary's ClassInfo.

Constructors

RegularF [MethodInfoFill]

ImplementsF [MethodInfoFill]

data MethodInfoFill

Mirrors ExternalLibrary's MethodInfo.

Constructors

CIF [ParameterFill] [Initializer] [StepFill]

MIF [ParameterFill] (NonEmpty StepFill)

externalLibCall :: [StepGroupFill] -> ExternalLibraryCall

Constructs an ExternalLibraryCall specification.

choiceStepsFill :: Int -> [StepFill] -> StepGroupFill

Corresponds to ExternalLibrary's choiceSteps. Provides the index of the steps that should be used for the current use case.

choiceStepFill :: Int -> StepFill -> StepGroupFill

Corresponds to ExternalLibrary's choiceStep. Provides the index of the step that should be used for the current use case.

mandatoryStepFill :: StepFill -> StepGroupFill

Corresponds to ExternalLibrary's mandatorySteps.

mandatoryStepsFill :: [StepFill] -> StepGroupFill

Corresponds to ExternalLibrary's mandatoryStep.

callStepFill :: FunctionIntFill -> StepFill

Corresponds to ExternalLibrary's callStep.

libCallFill :: [ArgumentFill] -> FunctionIntFill

Corresponds to any of ExternalLibrary's FunctionInterface constructors.

userDefinedArgFill :: CodeExpr -> ArgumentFill

Does not correspond to anything in ExternalLibrary. To be used when the presence of an argument is only a consequence of the use case.

basicArgFill :: CodeExpr -> ArgumentFill

Corresponds to ExternalLibrary's inlineArg, inlineNamedArg, preDefinedArg, and preDefinedNamedArg. Provides the CodeExpr for the argument's value.

functionArgFill :: [ParameterFill] -> StepFill -> ArgumentFill

Corresponds to ExternalLibrary's functionArg.

customObjArgFill :: [StateVariable] -> ClassInfoFill -> ArgumentFill

Corresponds to ExternalLibrary's customObjArg. Provides the list of state variables for the class that must be written in the calling program.

recordArgFill :: [CodeExpr] -> ArgumentFill

Corresponds to ExternalLibrary's recordArg. Provides the list of CodeExprs for the values of the fields that must be set by the calling program.

unnamedParamFill :: CodeVarChunk -> ParameterFill

Corresponds to ExternalLibrary's unnamedParam. Provides the CodeVarChunk representing the parameter.

unnamedParamPBVFill :: CodeVarChunk -> ParameterFill

Corresponds to ExternalLibrary's unnamedParam. Provides the CodeVarChunk representing the parameter. Specifies that the parameter is passed by value.

userDefinedParamFill :: CodeVarChunk -> ParameterFill

Does not correspond to anything in ExternalLibrary. To be used when the presence of a parameter is only a consequence of the use case.

customClassFill :: [MethodInfoFill] -> ClassInfoFill

Corresponds to ExternalLibrary's customClass.

implementationFill :: [MethodInfoFill] -> ClassInfoFill

Corresponds to ExternalLibrary's implementation.

constructorInfoFill :: [ParameterFill] -> [Initializer] -> [StepFill] -> MethodInfoFill

Corresponds to ExternalLibrary's constructorInfo. Provides Variable-Value pairs for variables initialized by the constructor.

methodInfoFill :: [ParameterFill] -> [StepFill] -> MethodInfoFill

Corresponds to ExternalLibrary's methodInfo.

appendCurrSolFill :: CodeVarChunk -> StepFill

Corresponds to ExternalLibrary's appendCurrSol. Provides the CodeVarChunk for the solution list.

populateSolListFill :: CodeVarChunk -> [StepFill]

Corresponds to ExternalLibrary's populateSolList. Provides the CodeVarChunk for the solution list.

assignArrayIndexFill :: CodeVarChunk -> [CodeExpr] -> StepFill

Corresponds to ExternalLibrary's assignArrayIndex. Provides the CodeVarChunk for the array variable. Provides the CodeExprs for the values to assign to each array index.

assignSolFromObjFill :: CodeVarChunk -> StepFill

Corresponds to ExternalLibrary's assignSolFromObj. Provides the CodeVarChunk for the variable that the solution should be assigned to.

initSolListFromArrayFill :: CodeVarChunk -> StepFill

Corresponds to ExternalLibrary's initSolListFromArray. Provides the CodeVarChunk for the solution list.

initSolListWithValFill :: CodeVarChunk -> CodeExpr -> StepFill

Corresponds to ExternalLibrary's initSolListWithVal. Provides the CodeVarChunk for the solution list and the CodeExpr for the initial element of the solution list

solveAndPopulateWhileFill :: FunctionIntFill -> CodeExpr -> FunctionIntFill -> CodeVarChunk -> StepFill

Corresponds to ExternalLibrary's solveAndPopulateWhile. Provides the CodeExpr for the upper bound in the while loop condition and the CodeVarChunk for the solution list.

returnExprListFill :: [CodeExpr] -> StepFill

Corresponds to ExternalLibrary's returnExprList. Provides the list of CodeExprs to return.

fixedStatementFill :: StepFill

Corresponds to ExternalLibrary's fixedReturn. No parameters because the statement is not use-case-dependent.

fixedStatementFill' :: CodeExpr -> StepFill

Corresponds to ExternalLibrary's fixedReturn'. use-case-specific a CodeExpr that parameterize the statement.

initSolWithValFill :: CodeVarChunk -> CodeExpr -> StepFill

Corresponds to ExternalLibrary's initSolWithVal. Provides the CodeVarChunk for one solution and one CodeExpr for the initial element of the solution list

data Lang

Various OO languages where code may be generated.

Constructors

Cpp

CSharp

Java

Python

Swift

Julia

class AuxiliarySym r => PackageSym r where

Members of this class must have all the information necessary for the AuxiliarySym in addition to information necessary to create a package.

Associated Types

type Package r

Methods

package :: ProgData -> [r (Auxiliary r)] -> r (Package r)

class AuxiliarySym r where

Members of this class must have a doxygen configuration, ReadMe file, sample input, omptimize doxygen document, information necessary for a makefile, auxiliary helper documents, and auxiliary from data documents.

Associated Types

type Auxiliary r

type AuxHelper r

Methods

doxConfig :: String -> GOOLState -> Verbosity -> r (Auxiliary r)

readMe :: ReadMeInfo -> r (Auxiliary r)

sampleInput :: ChunkDB -> DataDesc -> [Expr] -> r (Auxiliary r)

optimizeDox :: r (AuxHelper r)

makefile :: [FilePath] -> ImplementationType -> [Comments] -> GOOLState -> ProgData -> r (Auxiliary r)

auxHelperDoc :: r (AuxHelper r) -> Doc

auxFromData :: FilePath -> Doc -> r (Auxiliary r)

data AuxData

The underlying data type for auxiliary files in all renderers.

data PackData

The underlying data type for packages in all renderers.

data CodeChunk

Basic chunk representation in the code generation context. Contains a DefinedQuantityDict and the kind of code (variable or function).

data CodeVarChunk

Chunk representing a variable. The obv field represents the object containing this variable, if it is an object field.

data CodeFuncChunk

Chunk representing a function.

quantvar :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeVarChunk

Construct a CodeVarChunk from a Quantity.

quantfunc :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeFuncChunk

Construct a CodeFuncChunk from a Quantity.

ccObjVar :: CodeVarChunk -> CodeVarChunk -> CodeVarChunk

Combine an Object-type CodeChunk with another CodeChunk to create a new CodeChunk which represents a field of the first. ex. ccObjVar obj f = obj.f.

listToArray :: CodeVarChunk -> CodeVarChunk

field :: CodeExprC r => CodeVarChunk -> CodeVarChunk -> r

Constructs a CodeExpr representing the field of an actor

data ODEInfo

Structure to hold ODE information.

Constructors

ODEInfo

Fields

indepVar :: CodeVarChunk

Independent variable.

depVar :: CodeVarChunk

Dependent variable.

otherVars :: [CodeVarChunk]

Other variables in the ODE.

tInit :: CodeExpr

tFinal :: CodeExpr

initVal :: [CodeExpr]

Initial value of an ODE.

odeSyst :: [CodeExpr]

ODE equations.

odeOpts :: ODEOptions

Various options related to the ODE, including solution method, step size, initial value of a second order ODE, etc.

odeInfo :: CodeVarChunk -> CodeVarChunk -> [CodeVarChunk] -> CodeExpr -> CodeExpr -> [CodeExpr] -> [CodeExpr] -> ODEOptions -> ODEInfo

Basic ODEInfo constructor.

odeInfo' :: [CodeVarChunk] -> ODEOptions -> DifferentialModel -> InitialValueProblem -> ODEInfo

Create ODEInfo with Other variables, ODEOptions, DifferentialModel, and InitialValueProblem

data ODEOptions

Other parameters for solving the ODE numerically

Constructors

ODEOpts

Fields

solveMethod :: ODEMethod

Solution method.

absTol :: CodeExpr

Absolute tolerance.

relTol :: CodeExpr

Relative tolerance.

stepSize :: CodeExpr

Step size.

odeOptions :: ODEMethod -> CodeExpr -> CodeExpr -> CodeExpr -> ODEOptions

Basic ODEOptions constructor

data ODEMethod

Methods for solving ODEs. Includes Runge-Kutta 4-5, Backwards Differentiation Formula, or Adams' method.

Constructors

RK45

BDF

Adams

data ODELibPckg

Holds an ODE library package.

Constructors

ODELib

Fields

libName :: Name

Library name.

libVers :: Version

Version.

libSpec :: ExternalLibrary

Library specifications.

libCall :: ODEInfo -> ExternalLibraryCall

Library call.

libPath :: Maybe FilePath

Library path.

compatibleLangs :: [Lang]

Compatible OO languages.

mkODELib :: Name -> Version -> ExternalLibrary -> (ODEInfo -> ExternalLibraryCall) -> FilePath -> [Lang] -> ODELibPckg

Makes an ODELibPckg with the given name, ExternalLibrary specification, ExternalLibraryCall specification parameterized by an ODEInfo, local file path to the library, and list of compatible languages.

mkODELibNoPath :: Name -> Version -> ExternalLibrary -> (ODEInfo -> ExternalLibraryCall) -> [Lang] -> ODELibPckg

Makes an ODELibPckg with the given name, ExternalLibrary specification, ExternalLibraryCall specification parameterized by an ODEInfo, and list of compatible languages.

unPP :: PythonProject a -> a

unJP :: JavaProject a -> a

unCSP :: CSharpProject a -> a

unCPPP :: CppProject a -> a

unSP :: SwiftProject a -> a

unJLP :: JuliaProject a -> a

data NamedArgument

Any quantity can be a named argument (wrapper for DefinedQuantityDict), but with more of a focus on generating code arguments.

narg :: (Quantity q, MayHaveUnit q, Concept q) => q -> NamedArgument

Smart constructor for NamedArgument .

codeDQDs :: [DefinedQuantityDict]