Language.Drasil.Code.ExtLibImport

Description

Defines functions for simultaneous interpretation of ExternalLibrary and ExternalLibraryCall.

Documentation

data ExtLibState

State object used during interpretation of an ExternalLibrary and ExternalLibraryCall.

Constructors

ELS

Fields

_auxMods :: [Mod]

Additional modules that must be generated to use the library.

_defs :: [FuncStmt]

The defining statements for variables that must be pre-defined before being passed as arguments in an external library call.

_defined :: [Name]

The names of variables for which a defining statement has already been generated.

_steps :: [FuncStmt]

The statements corresponding to the external library use case.

_imports :: [String]

The imports required to use the external library.

_modExports :: [(Name, Name)]

An association list between library method/function names and the external library module that exports them.

auxMods :: Lens' ExtLibState [Mod]

defs :: Lens' ExtLibState [FuncStmt]

imports :: Lens' ExtLibState [String]

modExports :: Lens' ExtLibState [(Name, Name)]

steps :: Lens' ExtLibState [FuncStmt]

genExternalLibraryCall :: ExternalLibrary -> ExternalLibraryCall -> ExtLibState

Interprets an ExternalLibrary and ExternalLibraryCall and returns the resulting ExtLibState.