Language.Drasil.Code.Imperative.Build.AST

Documentation

type CommandFragment = MakeString

Used to build commands. Type synonym of a MakeString for clarity.

data BuildName

Type for holding the build name.

Constructors

BMain

BPackName

BPack BuildName

BWithExt BuildName Ext

data Ext

File extentions.

Constructors

CodeExt

OtherExt MakeString

data BuildDependencies

Builds may only require themselves or have a dependency.

Constructors

BcSource

BcSingle BuildName

data BuildConfig

Build configuration. In the function parameter, the first parameter is the list of inputs, 2nd parameter is the output file, 3rd parameter is additional name if needed. The two Maybe BuildNames are the output file and the additional name. Also holds the build dependencies.

Constructors

BuildConfig ([CommandFragment] -> CommandFragment -> CommandFragment -> [BuildCommand]) (Maybe BuildName) (Maybe BuildName) BuildDependencies

data RunType

Run commands as they are or through an interpreter.

Constructors

Standalone

Interpreter [CommandFragment]

data Runnable

Contains all the information needed to run a command.

Constructors

Runnable BuildName NameOpts RunType

data DocConfig

Configures a document based on dependencies and commands.

Constructors

DocConfig Dependencies [Command]

data NameOpts

Naming options. Includes a package separator and an option for including extensions.

Constructors

NameOpts

Fields

packSep :: String

includeExt :: Bool

nameOpts :: NameOpts

Default name options. Packages separately by "/" and includes extension.

type BuildCommand = [CommandFragment]

Build commands. Made up of CommandFragments.

type InterpreterCommand = String

Interpreter commands are made up of a String.

type InterpreterOption = String

Interpreter options are made up of a String.

asFragment :: String -> CommandFragment

Translates a String into a command fragment.

osClassDefault :: String -> String -> String -> CommandFragment

OS default variable (Windows). Calls mkWindowsVar.

buildAll :: ([CommandFragment] -> CommandFragment -> [BuildCommand]) -> BuildName -> Maybe BuildConfig

Constructor for a build configuration. No additional name included.

buildAllAdditionalName :: ([CommandFragment] -> CommandFragment -> CommandFragment -> [BuildCommand]) -> BuildName -> BuildName -> Maybe BuildConfig

Constructor for a build configuration with an additional name included.

buildSingle :: ([CommandFragment] -> CommandFragment -> [BuildCommand]) -> BuildName -> BuildName -> Maybe BuildConfig

Constructor for a build configuration. No additional name included, but takes in a single dependency.

nativeBinary :: Maybe Runnable

Default runnable information.

executable :: BuildName

Default target extension is ".exe".

sharedLibrary :: BuildName

Default library has the extentions ".dll, .dylib, .so".

interp :: BuildName -> NameOpts -> InterpreterCommand -> [InterpreterOption] -> Maybe Runnable

Constructor for a runnable command that goes through an interpreter.

interpMM :: InterpreterCommand -> Maybe Runnable

Constructs a runnable command that goes through an interpreter (for main module file).

mainModule :: BuildName

Main module.

mainModuleFile :: BuildName

Main module with a default code extension.

inCodePackage :: BuildName -> BuildName

Module for an in-code package.

withExt :: BuildName -> String -> BuildName

Adds an extension to a BuildName.

cCompiler :: CommandFragment

C compiler command fragment.

cppCompiler :: CommandFragment

C++ compiler command fragment.

doxygenDocConfig :: FilePath -> DocConfig

Helper for configuring doxygen documentation.