Language.Drasil.Code.Imperative.ConceptMatch

Description

Contains functions related to the choice of concept matches

Documentation

chooseConcept :: Choices -> State [Sentence] MatchedConceptMap

Concretizes the ConceptMatchMap in Choices to a MatchedConceptMap. Currently we don't have any Choices that would prevent a CodeConcept from being mapped, so we just take the head of the list of CodeConcepts The conceptMatchMap from choices is passed to chooseConept' internally, this way any codeconcept list can be matched to its appropiate UID

conceptToGOOL :: OOProg r => CodeConcept -> SValue r