Language.Drasil.Code.Imperative.Descriptions

Description

Defines description generators for common SCS functions, classes, and modules.

Documentation

modDesc :: GenState [Description] -> GenState Description

Returns a module description based on a list of descriptions of what is contained in the module.

unmodularDesc :: GenState Description

Returns description of what is contained in the module that is generated when the user chooses an Unmodular design. Module is described as either a program or library, depending on the user's choice of implementation type.

inputParametersDesc :: GenState [Description]

Returns description of what is contained in the Input Parameters module. If user chooses the Bundled input parameter, this module will include the structure for holding the input values. Does not include the structure if they choose Unbundled. This module includes the input-related functions.

inputConstructorDesc :: GenState Description

Returns a description of the input constructor, checking whether each possible method that may be called by the constructor is defined, and including it in the description if so.

inputFormatDesc :: GenState Description

Returns a description of what is contained in the Input Format module, if it exists.

derivedValuesDesc :: GenState Description

Returns a description of what is contained in the Derived Values module, if it exists.

inputConstraintsDesc :: GenState Description

Returns a description of what is contained in the Input Constraints module, if it exists.

constModDesc :: GenState Description

Returns a description of what is contained in the Constants module, if it exists.

outputFormatDesc :: GenState Description

Returns a description of what is contained in the Output Format module, if it exists.

inputClassDesc :: GenState Description

Returns a description for the generated function that stores inputs, if it exists. Checks whether explicit inputs, derived inputs, and constants are defined in the InputParameters class and includes each in the description if so.

constClassDesc :: GenState Description

Returns a description for the generated class that stores constants, if it exists. If no constants are defined in the Constants class, then it does not exist and an empty description is returned.

inFmtFuncDesc :: GenState Description

Returns a description for the generated function that reads input from a file, if it exists.

inConsFuncDesc :: GenState Description

Returns a description for the generated function that checks input constraints, if it exists.

dvFuncDesc :: GenState Description

Returns a description for the generated function that calculates derived inputs, if it exists.

calcModDesc :: Description

Description of the generated Calculations module.

woFuncDesc :: GenState Description

Returns description for generated output-printing function, if it exists.