Language.Drasil.Code.Imperative.Doxygen.Import

Documentation

makeDoxConfig :: ProjName -> GOOLState -> OptimizeChoice -> Verbosity -> Doc

Renders a Doxygen configuration file.

Sets the PROJECT_NAME field according to the given name.

Sets the INPUT field to the header files contained in the given GOOLState.

Sets the OPTIMIZE_OUTPUT_JAVA field according to the OptimizeChoice parameter.

Sets the QUIET field according to the given Verbosity.

yes :: Doc

Renderings of values commonly used in the configuration file.

no :: Doc

Renderings of values commonly used in the configuration file.