Language.Drasil.Code.Imperative.DrasilState

Documentation

type GenState = State DrasilState

Abbreviation used throughout generator.

data DrasilState

Private State, used to push these options around the generator.

Constructors

DrasilState

Fields

codeSpec :: CodeSpec

modular :: Modularity

implType :: ImplementationType

inStruct :: Structure

conStruct :: ConstantStructure

conRepr :: ConstantRepr

concMatches :: MatchedConceptMap

spaceMatches :: MatchedSpaces

onSfwrC :: ConstraintBehaviour

onPhysC :: ConstraintBehaviour

commented :: [Comments]

doxOutput :: Verbosity

date :: String

logName :: String

logKind :: [Logging]

auxiliaries :: [AuxFile]

sampleData :: [Expr]

dsICNames :: InternalConcept -> Name

modules :: [Mod]

extLibNames :: [(Name, Version)]

extLibMap :: ExtLibMap

libPaths :: [FilePath]

eMap :: ModExportMap

libEMap :: ModExportMap

clsMap :: ClassDefinitionMap

defSet :: Set Name

getVal :: Int

currentModule :: String

currentClass :: String

_designLog :: Doc

_loggedSpaces :: [(Space, CodeType)]

currentScope :: ScopeType

designLog :: Lens' DrasilState Doc

type MatchedSpaces = Space -> GenState CodeType

Type for the mapping between Spaces and CodeTypes.

type ModExportMap = Map String String

Variable/function name maps to module name.

type ClassDefinitionMap = Map String String

Variable/function name maps to class name.

data ScopeType

Variable scope

Constructors

Local

Global

MainFn

modExportMap :: OldCodeSpec -> Choices -> [Mod] -> ModExportMap

Builds the module export map, mapping each function and state variable name in the generated code to the name of the generated module that exports it.

clsDefMap :: OldCodeSpec -> Choices -> [Mod] -> ClassDefinitionMap

Builds the class definition map, mapping each generated method and state variable name to the name of the generated class where it is defined.

addToDesignLog :: Space -> CodeType -> Doc -> DrasilState -> DrasilState

Adds a message to the design log if the given Space-CodeType match has not already been logged.

addLoggedSpace :: Space -> CodeType -> DrasilState -> DrasilState

Adds a Space-CodeType pair to the loggedSpaces list in DrasilState to prevent a duplicate log from being generated for that Space-CodeType pair.

genICName :: InternalConcept -> GenState Name

Get InternalConcept name using DrasilState