Language.Drasil.Code.Imperative.FunctionCalls

Documentation

getAllInputCalls :: OOProg r => GenState [MSStatement r]

Generates calls to all of the input-related functions. First is the call to the function for reading inputs, then the function for calculating derived inputs, then the function for checking input constraints.

getInputCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for reading inputs from a file.

getDerivedCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for calculating derived inputs.

getConstraintCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for checking constraints on the input.

getCalcCall :: OOProg r => CodeDefinition -> GenState (Maybe (MSStatement r))

Generates a call to a calculation function, given the CodeDefinition for the value being calculated.

getOutputCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for printing outputs.