Language.Drasil.Code.Imperative.GOOL.ClassInterface

Description

Defines a package extension for GOOL, with functions for pairing a GOOL program with auxiliary, non-source-code files.

Documentation

data SoftwareDossierState

makeSds :: [FilePath] -> [FilePath] -> Maybe FilePath -> SoftwareDossierState

headers :: Lens' SoftwareDossierState [FilePath]

sources :: Lens' SoftwareDossierState [FilePath]

mainMod :: Lens' SoftwareDossierState (Maybe FilePath)

class AuxiliarySym r where

Members of this class must have a doxygen configuration, ReadMe file, omptimize doxygen document, information necessary for a makefile, and auxiliary helper documents

Methods

doxConfig :: String -> SoftwareDossierState -> Verbosity -> r FileAndContents

readMe :: ReadMeInfo -> r FileAndContents

optimizeDox :: r Doc

makefile :: [FilePath] -> ImplementationType -> [Comments] -> SoftwareDossierState -> ProgData -> r FileAndContents

auxHelperDoc :: r Doc -> Doc

sampleInput :: Applicative r => PrintingInformation -> DataDesc -> [Expr] -> r FileAndContents

auxFromData :: Applicative r => FilePath -> Doc -> r FileAndContents