Language.Drasil.Code.Imperative.GOOL.Data

Description

Defines the underlying data types used in the package extension.

Documentation

data AuxData

The underlying data type for auxiliary files in all renderers.

ad :: FilePath -> Doc -> AuxData

Constructor for auxiliary files.

data PackData

The underlying data type for packages in all renderers.

packD :: ProgData -> [AuxData] -> PackData

Constructor for package data.