Language.Drasil.Code.Imperative.GenODE

Documentation

chooseODELib :: Lang -> Maybe ODE -> State [Sentence] ODEGenInfo

Chooses the first ODELibPckg from the list specified by the user that is compatible with the current target Lang. Interprets the ExternalLibrary and ExternalLibraryCall for the selected ODELibPckg by concretizing the ExternalLibraryCall with each of the ODEInfos The internal helper chooseODELib' keeps a read only preference list and a currently considered preference list (which can change), this facilitates the firstChoiceODELib check.