Language.Drasil.Code.Imperative.Generator

Description

Defines generation functions for SCS code packages.

Documentation

generator :: Lang -> String -> [Expr] -> Choices -> CodeSpec -> DrasilState

Initializes the generator's DrasilState. Space parameter is a string representing the date. \['Expr'\] parameter is the sample input values provided by the user.

generateCode :: (OOProg progRepr, PackageSym packRepr) => Lang -> (progRepr (Program progRepr) -> ProgData) -> (packRepr (Package packRepr) -> PackData) -> DrasilState -> IO ()

Generates a package with the given DrasilState. The passed un-representation functions determine which target language the package will be generated in.

generateCodeProc :: (ProcProg progRepr, PackageSym packRepr) => Lang -> (progRepr (Program progRepr) -> ProgData) -> (packRepr (Package packRepr) -> PackData) -> DrasilState -> IO ()

Generates a package with the given DrasilState. The passed un-representation functions determine which target language the package will be generated in.