Language.Drasil.Code.Imperative.Modules

Documentation

genMain :: OOProg r => GenState (SFile r)

genMainFunc :: OOProg r => GenState (Maybe (SMethod r))

chooseInModule :: OOProg r => InputModule -> GenState [SFile r]

genInputClass :: OOProg r => ClassType -> GenState (Maybe (SClass r))

Returns Nothing if no inputs or constants are mapped to InputParameters in the class definition map. If any inputs or constants are defined in InputParameters, this generates the InputParameters class containing the inputs and constants as state variables. If the InputParameters constructor is also exported, then the generated class also contains the input-related functions as private methods

genInputDerived :: OOProg r => ScopeTag -> GenState (Maybe (SMethod r))

genInputConstraints :: OOProg r => ScopeTag -> GenState (Maybe (SMethod r))

genInputFormat :: OOProg r => ScopeTag -> GenState (Maybe (SMethod r))

genConstMod :: OOProg r => GenState [SFile r]

genConstClass :: OOProg r => ClassType -> GenState (Maybe (SClass r))

genCalcMod :: OOProg r => GenState (SFile r)

genCalcFunc :: OOProg r => CodeDefinition -> GenState (SMethod r)

genOutputMod :: OOProg r => GenState [SFile r]

genOutputFormat :: OOProg r => GenState (Maybe (SMethod r))

genSampleInput :: AuxiliarySym r => GenState (Maybe (r (Auxiliary r)))