Language.Drasil.Code.Imperative.Parameters

Documentation

getInConstructorParams :: GenState [CodeVarChunk]

Since the input constructor calls the three input-related methods, the parameters to the constructor are the parameters to the three methods, except excluding any of variables that are state variables in the class, since they are already in scope. If InputParameters is not in the definition list, then the default constructor is used, which takes no parameters.

getInputFormatIns :: GenState [CodeVarChunk]

The inputs to the function for reading inputs are the input file name.

getInputFormatOuts :: GenState [CodeVarChunk]

The outputs from the function for reading inputs are the inputs.

getDerivedIns :: GenState [CodeVarChunk]

The inputs to the function for calculating derived inputs are any variables used in the equations for the derived inputs.

getDerivedOuts :: GenState [CodeVarChunk]

The outputs from the function for calculating derived inputs are the derived inputs.

getConstraintParams :: GenState [CodeVarChunk]

The parameters to the function for checking constraints on the inputs are any inputs with constraints, and any variables used in the expressions of the constraints.

getCalcParams :: CodeDefinition -> GenState [CodeVarChunk]

The parameters to a calculation function are any variables used in the expression representing the calculation.

getOutputParams :: GenState [CodeVarChunk]

The parameters to the function for printing outputs are the outputs.