Language.Drasil.Code.Imperative.ReadInput

Description

Defines functions for reading values from a file corresponding to a DataDesc

Documentation

sampleInputDD :: [CodeVarChunk] -> DataDesc'

Defines the DataDesc for the file containing a sample data set, which a user must supply if they want to generate a sample input file.

readWithDataDesc :: FilePath -> DataDesc' -> IO [Expr]

Reads data from a file and converts the values to Exprs. The file must be formatted according to the DataDesc' passed as a parameter.