Language.Drasil.Code.Imperative.ReadMe.Import

Description

This module defines structures and functions for handling README files, which are used as auxiliary documentation files for the system.

Documentation

data ReadMeInfo

Holds all information needed to create a README file.

Constructors

ReadMeInfo

Fields

langName :: LangAbbrev

langVersion :: LangVers

invalidOS :: Maybe String

implementType :: ImplementationType

extLibNV :: [(Name, Version)]

extLibFP :: [FilePath]

contributors :: [Contributor]

configFP :: [FilePath]

caseName :: CaseName

examplePurpose :: ExamplePurpose

exampleDescr :: ExampleDescr

exampleMotivation :: ExampleMotivation

exampleScope :: ExampleScope

folderNum :: Int

inputOutput :: (InFile, OutFile)