Language.Drasil.CodeSpec

Description

Defines the CodeSpec structure and related functions.

Documentation

type Input = CodeVarChunk

Program input.

type Output = CodeVarChunk

Program output.

type Const = CodeDefinition

Constants in the problem.

type Derived = CodeDefinition

Derived inputs.

type Def = CodeDefinition

Mathematical definition.

data CodeSpec where

Code specifications. Holds information needed to generate code.

Constructors

CodeSpec

Fields

:: HasName a

=> { pName :: Name

Program name.

, authors :: [a]

Authors.

, inputs :: [Input]

All inputs.

, extInputs :: [Input]

Explicit inputs (values to be supplied by a file).

, derivedInputs :: [Derived]

Derived inputs (each calculated from explicit inputs in a single step).

, outputs :: [Output]

All outputs.

, configFiles :: [FilePath]

List of files that must be in same directory for running the executable.

, execOrder :: [Def]

Mathematical definitions, ordered so that they form a path from inputs to outputs.

, cMap :: ConstraintCEMap

Map from UIDs to constraints for all constrained chunks used in the problem.

, constants :: [Const]

List of all constants used in the problem.

, constMap :: ConstantMap

Map containing all constants used in the problem.

, mods :: [Mod]

Additional modules required in the generated code, which Drasil cannot yet automatically define.

, sysinfodb :: ChunkDB

The database of all chunks used in the problem.

} -> CodeSpec

type ConstantMap = Map UID CodeDefinition

Maps constants to their respective CodeDefinition.

assocToMap :: HasUID a => [a] -> Map UID a

Converts a list of chunks that have UIDs to a Map from UID to the associated chunk.

getODE :: [ExtLib] -> Maybe ODE

Get ODE from ExtLib

mapODE :: Maybe ODE -> [CodeDefinition]

Maps ODE to their respective CodeDefinition.

codeSpec :: SystemInformation -> Choices -> [Mod] -> CodeSpec

Defines a CodeSpec based on the SystemInformation, Choices, and Mods defined by the user.

asVC :: Func -> QuantityDict

Convert a Func to an implementation-stage QuantityDict representing the function.

funcUID :: Func -> UID

Get a UID of a chunk corresponding to a Func.

getDerivedInputs :: [DataDefinition] -> [Input] -> [Const] -> ChunkDB -> [SimpleQDef]

Determines the derived inputs, which can be immediately calculated from the knowns (inputs and constants). If there are DDs, the derived inputs will come from those. If there are none, then the QDefinitions are used instead.

type Known = CodeVarChunk

Known values.

type Need = CodeVarChunk

Calculated values.

getExecOrder :: [Def] -> [Known] -> [Need] -> ChunkDB -> [Def]

Orders a list of definitions such that they form a path between Known values and values that Need to be calculated.

getConstraints :: HasUID c => ConstraintCEMap -> [c] -> [ConstraintCE]

Get a list of Constraints for a list of CodeChunks.

constraintvars :: ConstraintCE -> ChunkDB -> [CodeChunk]

Get a list of CodeChunks from a constraint.