Language.Drasil.GOOL

Description

Re-export code-related smart constructors for external code writing and generation.

Documentation

class AuxiliarySym r where

Members of this class must have a doxygen configuration, ReadMe file, omptimize doxygen document, information necessary for a makefile, and auxiliary helper documents

Methods

doxConfig :: String -> SoftwareDossierState -> Verbosity -> r FileAndContents

readMe :: ReadMeInfo -> r FileAndContents

optimizeDox :: r Doc

makefile :: [FilePath] -> ImplementationType -> [Comments] -> SoftwareDossierState -> ProgData -> r FileAndContents

auxHelperDoc :: r Doc -> Doc

package :: Monad r => ProgData -> [r FileAndContents] -> r (PackageData ProgData)

data FileAndContents

The underlying data type for auxiliary files in all renderers.

hasPathAndDocToFileAndContents :: HasPathAndDoc a Doc => a -> FileAndContents

data PackageData a

The underlying data type for packages in all renderers.

pattern PackageData :: a -> [FileAndContents] -> PackageData a

unPP :: PythonProject a -> a

unJP :: JavaProject a -> a

unCSP :: CSharpProject a -> a

unCPPP :: CppProject a -> a

unSP :: SwiftProject a -> a

unJLP :: JuliaProject a -> a