Language.Drasil.ICOSolutionSearch

Documentation

type Def = CodeDefinition

Mathematical definition.

type Known = CodeVarChunk

Known values.

type Need = CodeVarChunk

Calculated values.

getExecOrder :: [Def] -> [Known] -> [Need] -> ChunkDB -> [Def]

Orders a list of definitions such that they form a path between Known values and values that Need to be calculated.