Language.Drasil.Mod

Description

Defines an AST for defining Modules.

Documentation

data Class

Holds information needed to define a class.

Constructors

ClassDef

Fields

className :: Name

implements :: Maybe Name

classDesc :: Description

stateVars :: [StateVariable]

constructors :: [Func]

methods :: [Func]

data StateVariable

State variables hold attach a VisibilityTag to a CodeVarChunk.

Constructors

SV

Fields

svVisibility :: VisibilityTag

stVar :: CodeVarChunk

data Func

Holds a function definition or function data.

Constructors

FDef FuncDef

FData FuncData

data FuncData where

Function data. Holds a name, description, and pieces of data with its own description.

Constructors

FuncData :: Name -> Description -> DataDesc -> FuncData

data FuncDef where

Defines a function.

Constructors

FuncDef :: Name -> Description -> [ParameterChunk] -> Space -> Maybe Description -> [FuncStmt] -> FuncDef
Parameters are: Name, description, parameters, return type, return description, statements.

CtorDef :: Name -> Description -> [ParameterChunk] -> [Initializer] -> [FuncStmt] -> FuncDef

data FuncStmt where

Constructors

FAsg :: CodeVarChunk -> CodeExpr -> FuncStmt

FAsgIndex :: CodeVarChunk -> Integer -> CodeExpr -> FuncStmt

FFor :: CodeVarChunk -> CodeExpr -> CodeExpr -> CodeExpr -> [FuncStmt] -> FuncStmt
For-loop; Variable, Start, Stop, Step, Body.

FForEach :: CodeVarChunk -> CodeExpr -> [FuncStmt] -> FuncStmt

FWhile :: CodeExpr -> [FuncStmt] -> FuncStmt

FCond :: CodeExpr -> [FuncStmt] -> [FuncStmt] -> FuncStmt

FRet :: CodeExpr -> FuncStmt

FThrow :: String -> FuncStmt

FTry :: [FuncStmt] -> [FuncStmt] -> FuncStmt

FContinue :: FuncStmt

FDecDef :: CodeVarChunk -> CodeExpr -> FuncStmt

FFuncDef :: CodeFuncChunk -> [ParameterChunk] -> [FuncStmt] -> FuncStmt

FVal :: CodeExpr -> FuncStmt

FMulti :: [FuncStmt] -> FuncStmt

FAppend :: CodeExpr -> CodeExpr -> FuncStmt

type Initializer = (CodeVarChunk, CodeExpr)

Variable-value pair.

data Mod

Holds module information.

Constructors

Mod Name Description [Import] [Class] [Func]

type Name = String

Type synonym for clarity.

type Version = String

Type synonym for clarity.

type Description = String

Type synonym for clarity.

type Import = String

Type synonym for clarity.

($:=) :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> FuncStmt

Define an assignment statement.

pubStateVar :: CodeVarChunk -> StateVariable

Define a public state variable based on the given CodeVarChunk.

privStateVar :: CodeVarChunk -> StateVariable

Define a private state variable based on the given CodeVarChunk.

classDef :: Name -> Description -> [StateVariable] -> [Func] -> [Func] -> Class

Define a class with the given Name, Description, state variables, and methods.

classImplements :: Name -> Name -> Description -> [StateVariable] -> [Func] -> [Func] -> Class

Define a class that implements an interface. 1st Name is class name, 2nd is interface name.

ctorDef :: Name -> Description -> [ParameterChunk] -> [Initializer] -> [FuncStmt] -> Func

Define a constructor, with the given name, description, parameters, initializers (variable-value pairs), and FuncStmts for the body.

ffor :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> [FuncStmt] -> FuncStmt

Define a for-loop. Quantity is for the iteration variable, CodeExpr is the upper bound at that variable (the variable will start with a value of 0). [FuncStmt] is for the loop body.

fforRange :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> CodeExpr -> CodeExpr -> [FuncStmt] -> FuncStmt

Define a for-loop. Quantity is for the iteration variable, and 3 CodeExprs for the start, stop, step numbers. [FuncStmt] is for the loop body.

fDecDef :: (Quantity c, MayHaveUnit c, Concept c) => c -> CodeExpr -> FuncStmt

Define a declare-define statement.

fname :: Func -> Name

Gets the name of a function.

fstdecl :: ChunkDB -> [FuncStmt] -> [CodeVarChunk]

Returns the list of CodeVarChunks that are used in the list of FuncStmts but are not declared in any of the FuncStmts.

funcData :: Name -> Description -> DataDesc -> Func

Define a function that reads data from a file, according to the given DataDesc.

funcDef :: (Quantity c, MayHaveUnit c, Concept c) => Name -> Description -> [c] -> Space -> Maybe Description -> [FuncStmt] -> Func

Define a function by providing the FuncStmts for its body. Other parameters are function name, description, list of parameters, space of the returned value, and description of the returned value.

funcDefParams :: Name -> Description -> [ParameterChunk] -> Space -> Maybe Description -> [FuncStmt] -> Func

Like funcDef but uses ParameterChunks to represent the parameters.

packmod :: Name -> Description -> [Class] -> [Func] -> Mod

Define a Mod with the given Name, Description, Classes, and Functions.

packmodRequires :: Name -> Description -> [Import] -> [Class] -> [Func] -> Mod

Define a Mod that requires some library imports, with the given Name, Description, Classes, and Functions.