Language.Drasil.Code.Imperative.FunctionCalls

Documentation

genAllInputCalls :: OOProg r => GenState [MSStatement r]

Generates calls to all of the input-related functions. First is the call to the function for reading inputs, then the function for calculating derived inputs, then the function for checking input constraints.

genAllInputCallsProc :: SharedProg r => GenState [MSStatement r]

Generates calls to all of the input-related functions. First is the call to the function for reading inputs, then the function for calculating derived inputs, then the function for checking input constraints.

genInputCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for reading inputs from a file.

genInputCallProc :: SharedProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for reading inputs from a file.

genDerivedCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for calculating derived inputs.

genDerivedCallProc :: SharedProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for calculating derived inputs.

genConstraintCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for checking constraints on the input.

genConstraintCallProc :: SharedProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for checking constraints on the input.

genCalcCall :: OOProg r => CodeDefinition -> GenState (Maybe (MSStatement r))

Generates a call to a calculation function, given the CodeDefinition for the value being calculated.

genCalcCallProc :: SharedProg r => CodeDefinition -> GenState (Maybe (MSStatement r))

Generates a call to a calculation function, given the CodeDefinition for the value being calculated.

genOutputCall :: OOProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for printing outputs.

genOutputCallProc :: SharedProg r => GenState (Maybe (MSStatement r))

Generates a call to the function for printing outputs.