Language.Drasil.Code.Imperative.GOOL.ClassInterface

Description

Defines a package extension for GOOL, with functions for pairing a GOOL program with auxiliary, non-source-code files.

Documentation

class AuxiliarySym r => PackageSym r where

Members of this class must have all the information necessary for the AuxiliarySym in addition to information necessary to create a package.

Associated Types

type Package r

Methods

package :: ProgData -> [r (Auxiliary r)] -> r (Package r)

class AuxiliarySym r where

Members of this class must have a doxygen configuration, ReadMe file, sample input, omptimize doxygen document, information necessary for a makefile, auxiliary helper documents, and auxiliary from data documents.

Associated Types

type Auxiliary r

type AuxHelper r

Methods

doxConfig :: String -> GOOLState -> Verbosity -> r (Auxiliary r)

readMe :: ReadMeInfo -> r (Auxiliary r)

sampleInput :: ChunkDB -> DataDesc -> [Expr] -> r (Auxiliary r)

optimizeDox :: r (AuxHelper r)

makefile :: [FilePath] -> ImplementationType -> [Comments] -> GOOLState -> ProgData -> r (Auxiliary r)

auxHelperDoc :: r (AuxHelper r) -> Doc

auxFromData :: FilePath -> Doc -> r (Auxiliary r)