Language.Drasil.CodeSpec

Description

Defines the CodeSpec structure and related functions.

Documentation

type Input = CodeVarChunk

Program input.

type Output = CodeVarChunk

Program output.

type Const = CodeDefinition

Constants in the problem.

type Derived = CodeDefinition

Derived inputs.

type Def = CodeDefinition

Mathematical definition.

type ConstantMap = Map UID CodeDefinition

Maps constants to their respective CodeDefinition.

data OldCodeSpec

Old Code specifications. Holds information needed to generate code.

Constructors

OldCodeSpec

Fields

_pName :: Name

Program name.

_authors :: People

Authors.

_inputs :: [Input]

All inputs.

_extInputs :: [Input]

Explicit inputs (values to be supplied by a file).

_derivedInputs :: [Derived]

Derived inputs (each calculated from explicit inputs in a single step).

_outputs :: [Output]

All outputs.

_configFiles :: [FilePath]

List of files that must be in same directory for running the executable.

_execOrder :: [Def]

Mathematical definitions, ordered so that they form a path from inputs to outputs.

_cMap :: ConstraintCEMap

Map from UIDs to constraints for all constrained chunks used in the problem.

_constants :: [Const]

List of all constants used in the problem.

_constMap :: ConstantMap

Map containing all constants used in the problem.

_mods :: [Mod]

Additional modules required in the generated code, which Drasil cannot yet automatically define.

_systemdb :: ChunkDB

The database of all chunks used in the problem.

class HasOldCodeSpec c where

Minimal complete definition

oldCodeSpec

Methods

oldCodeSpec :: Lens' c OldCodeSpec

authorsO :: Lens' c People

cMapO :: Lens' c ConstraintCEMap

configFilesO :: Lens' c [FilePath]

constMapO :: Lens' c ConstantMap

constantsO :: Lens' c [Const]

derivedInputsO :: Lens' c [Derived]

execOrderO :: Lens' c [Def]

extInputsO :: Lens' c [Input]

inputsO :: Lens' c [Input]

modsO :: Lens' c [Mod]

outputsO :: Lens' c [Output]

pNameO :: Lens' c Name

systemdbO :: Lens' c ChunkDB

data CodeSpec

New Code Specification. Holds system information and a reference to OldCodeSpec.

Constructors

CS

Fields

_system' :: System

_oldCode :: OldCodeSpec

system' :: Lens' CodeSpec System

oldCode :: Lens' CodeSpec OldCodeSpec

assocToMap :: HasUID a => [a] -> Map UID a

Converts a list of chunks that have UIDs to a Map from UID to the associated chunk.

getODE :: [ExtLib] -> Maybe ODE

Get ODE from ExtLib

mapODE :: Maybe ODE -> [CodeDefinition]

Maps ODE to their respective CodeDefinition.

codeSpec :: System -> Choices -> [Mod] -> CodeSpec

Creates a CodeSpec using the provided System, Choices, and Mods. The CodeSpec consists of the system information and a corresponding OldCodeSpec.

oldcodeSpec :: System -> Choices -> [Mod] -> OldCodeSpec

Generates an OldCodeSpec from System, Choices, and a list of Mods. This function extracts various components (e.g., inputs, outputs, constraints, etc.) from System to populate the OldCodeSpec structure.

asVC :: Func -> DefinedQuantityDict

Convert a Func to an implementation-stage DefinedQuantityDict representing the function.

funcUID :: Func -> UID

Get a UID of a chunk corresponding to a Func.

getDerivedInputs :: [DataDefinition] -> [Input] -> [Const] -> ChunkDB -> [SimpleQDef]

Determines the derived inputs, which can be immediately calculated from the knowns (inputs and constants). If there are DDs, the derived inputs will come from those. If there are none, then the QDefinitions are used instead.

type Known = CodeVarChunk

Known values.

type Need = CodeVarChunk

Calculated values.

getExecOrder :: [Def] -> [Known] -> [Need] -> ChunkDB -> [Def]

Orders a list of definitions such that they form a path between Known values and values that Need to be calculated.

getConstraints :: HasUID c => ConstraintCEMap -> [c] -> [ConstraintCE]

Get a list of Constraints for a list of CodeChunks.

constraintvars :: ConstraintCE -> ChunkDB -> [CodeChunk]

Get a list of CodeChunks from a constraint.