Language.Drasil.GOOL

Description

Re-export code-related smart constructors for external code writing and generation.

Documentation

class AuxiliarySym r => PackageSym r where

Members of this class must have all the information necessary for the AuxiliarySym in addition to information necessary to create a package.

Associated Types

type Package r

Methods

package :: ProgData -> [r (Auxiliary r)] -> r (Package r)

class AuxiliarySym r where

Members of this class must have a doxygen configuration, ReadMe file, sample input, omptimize doxygen document, information necessary for a makefile, auxiliary helper documents, and auxiliary from data documents.

Associated Types

type Auxiliary r

type AuxHelper r

Methods

doxConfig :: String -> GOOLState -> Verbosity -> r (Auxiliary r)

readMe :: ReadMeInfo -> r (Auxiliary r)

sampleInput :: PrintingInformation -> DataDesc -> [Expr] -> r (Auxiliary r)

optimizeDox :: r (AuxHelper r)

makefile :: [FilePath] -> ImplementationType -> [Comments] -> GOOLState -> ProgData -> r (Auxiliary r)

auxHelperDoc :: r (AuxHelper r) -> Doc

auxFromData :: FilePath -> Doc -> r (Auxiliary r)

data AuxData

The underlying data type for auxiliary files in all renderers.

data PackData

The underlying data type for packages in all renderers.

unPP :: PythonProject a -> a

unJP :: JavaProject a -> a

unCSP :: CSharpProject a -> a

unCPPP :: CppProject a -> a

unSP :: SwiftProject a -> a

unJLP :: JuliaProject a -> a